Rudely extend fluent to nanosecond precision

+++ lib/fluent/parser.rb
@@ -75,7 +75,7 @@ module Fluent
           return @cache2_time
         else
           begin
-            time = @parser.call(value).to_i
+          time = @parser.call(value).to_f
           rescue => e
             raise ParserError, "invalid time format: value = #{value}, error_class = #{e.class.name}, error = #{e.message}"
           end
